+++ devel/combo.R
@@ -172,9 +172,9 @@ myStopping3 <- new("StoppingTargetBiomarker",
 
 ## you can either write this:
 myStopping <- new("StoppingAny",
-                  stopList=
+                  stop_list=
                   list(new("StoppingAll",
-                           stopList=
+                           stop_list=
                            list(myStopping1,
                                 myStopping3)),
                        myStopping2))