+++ examples/Rules-method-size-CohortSizeMax.R
@@ -42,7 +42,7 @@ doseRecommendation <- nextBest(myNextBest,
 # Rule for having cohort of size 1 for doses <30
 #      and having cohort of size 3 for doses >=30
 mySize1 <- CohortSizeRange(intervals = c(0, 10),
-                           cohortSize = c(1, 3))
+                           cohort_size = c(1, 3))
 
 # Rule for having cohort of size 1 until no DLT were observed
 #      and having cohort of size 3 as soon as 1 DLT is observed