+++ examples/Rules-method-stopTrialCIMaxGainSamples.R
@@ -1,5 +1,5 @@
 # nolint start
-##define the stopping rules based on the 'StoppingGstarCIRatio' class
+##define the stopping rules based on the 'StoppingMaxGainCIRatio' class
 ##Using both DLE and efficacy responses
 ## we need a data object with doses >= 1:
 data <-DataDual(x=c(25,50,25,50,75,300,250,150),
@@ -21,9 +21,8 @@ options<-McmcOptions(burnin=10,step=2,samples=50)
 DLEsamples<-mcmc(data,DLEmodel,options)
 Effsamples<-mcmc(data,Effmodel,options)
 
-##define the 'StoppingGstarCIRatio' class
-myStopping <- StoppingGstarCIRatio(targetRatio=5,
-                                          targetEndOfTrial=0.3)
+##define the 'StoppingMaxGainCIRatio' class
+myStopping <- StoppingMaxGainCIRatio(target_ratio = 5, prob_target = 0.3)
 ##Find the next Recommend dose using the nextBest method (plesae refer to nextbest examples)
 mynextbest <- NextBestMaxGainSamples(
   prob_target_drt = 0.35,